Clean up of files and folders structure
