--- lib/api/groups.json
@@ -1,1 +1,0 @@
-[{"id":"01abcde234","name":"Group A","private":false,"contacts":[{"id":"01abcde235","name":"John Doe","email":"john.doe@example.com"},{"id":"01abcde236","name":"Jane Doe","email":"jane.doe@example.com"},{"id":"01abcde237","name":"Bob Smith","email":"bob.smith@example.com"},{"id":"01abcde238","name":"Alice Smith","email":"alice.smith@example.com"},{"id":"01abcde239","name":"David Brown","email":"david.brown@example.com"}],"owners":[{"id":"01abcde235","name":"John Doe","email":"john.doe@example.com"},{"id":"01abcde236","name":"Jane Doe","email":"jane.doe@example.com"}]},{"id":"01abcde24a","name":"Group B","private":true,"contacts":[{"id":"01abcde24b","name":"Bob Johnson","email":"bob.johnson@example.com"},{"id":"01abcde24c","name":"Alice Johnson","email":"alice.johnson@example.com"},{"id":"01abcde24d","name":"Jane Smith","email":"jane.smith@example.com"},{"id":"01abcde24e","name":"John Smith","email":"john.smith@example.com"},{"id":"01abcde24f","name":"Susan Brown","email":"susan.brown@example.com"}],"owners":[{"id":"01abcde24b","name":"Bob Johnson","email":"bob.johnson@example.com"},{"id":"01abcde24c","name":"Alice Johnson","email":"alice.johnson@example.com"}]}]
