+++ lib/src/model/group.dart
@@ -1,8 +1,36 @@
-/*import 'package:flutter/widgets.dart';
-import 'user.dart';
-import 'package:uuid/uuid.dart';
+//import 'package:flutter/widgets.dart';
+//import 'user.dart';
+//import 'package:uuid/uuid.dart';
 
-class ContactGroupModel extends ChangeNotifier {
+class Group {
+  String name;
+  String type;
+
+  Group({required this.name, required this.type});
+
+  factory Group.fromJson(Map<String, dynamic> json) {
+    return Group(
+      name: json['name'] as String,
+      type: json['type'] as String,
+    );
+  }
+}
+
+class GroupContacts {
+  final String name;
+  final String contactId;
+
+  GroupContacts({required this.name, required this.contactId});
+
+  factory GroupContacts.fromJson(Map<String, dynamic> json) {
+    return GroupContacts(
+      name: json['name'] as String,
+      contactId: json['contactId'] as String,
+    );
+  }
+}
+
+/*class ContactGroupModel extends ChangeNotifier {
     final String id;
     final String name;
     final bool private;
