+++ lib/src/model/place.dart
@@ -1,1 +1,1 @@
-//class to model places to store them instead of querry google apis every time.
+//class to model places to store them instead of query google apis every time.
